Application.App

Description

Documentation

buildInitialState :: Config -> IO AppState

Wrap the config file entries in a nonempty list.

tcmApp :: App AppState e ResourceName

For this application only event handling, drawing and some attributes need to be implemented. The rest are default implementations.

data AppState

The state of the application. It will guarantee that the list of items contains at least one element and keep track of a cursor position.

runApp :: IO ()

Run the app by loading items from the config file, turning them into an initial state and then delegating to Brick. Unless a yaml parsing exception is printed to facilitate config file debugging, there is no failure condition and the program can exit successfully. This situation may change after permission issues are considered.