Domain.State

Description

Documentation

data AppState

The state of the application. It will guarantee that the list of items contains at least one element and keep track of a cursor position.

Constructors

MkAppState (NonEmptyCursor ConfigItem)

type NextAppState = EventM ResourceName (Next AppState)

When reacting to an event the result needs to be of type EventM. Though not entirely accurate the synonym here is trying to communicate that, in essence, a new state of the application is created.

data ResourceName

An dummy type used only as a parameter of EventM