Infrastructure.FileModification

Description

Documentation

modifyFile :: FilePath -> (Content -> Content) -> IO ()

Given a filepath and a function which takes old content to new content, will apply the function to the file content and, consequently modify the file on disk.

data Content

A type representing file content.

Constructors

MkContent Text