Infrastructure.Util

Description

Documentation

changeNthElement :: Int -> (t -> t) -> [t] -> [t]

Given an index, a function and a list it applies the function to the element at that index.

changeNthElementNonEmpty :: Int -> (t -> t) -> NonEmpty t -> NonEmpty t

Given an index, a function and a NonEmpty list it applies the function to the element at that index.

elementAfter :: Eq t => t -> [t] -> t

Finds the first element equal to the input element inside of a list and returns the element after it (one index up).

elementBefore :: Eq t => t -> [t] -> t

Finds the first element equal to the input element inside of a list and returns the element before it (one index down).