UserInterface.Render

Description

Documentation

drawTCM :: AppState -> [Widget ResourceName]

The rendering consists of a single layer, each line consists of an items title and current value. The selected line is rendered boldface, the selected value, additionally, has a separate color.

selectionStyling :: Attr

Define the styling to be applied to selected items.

valueStyling :: Attr

Define the styling to be applied to the value of a selected line.